Beter separation of project properties and packages, between tests and source code. Adjust visibility of some types (#264)

+++ src/PaperMalKing.Database.CompiledModels/ShikiUserEntityType.cs
@@ -51,7 +51,7 @@ namespace PaperMalKing.Database.CompiledModels
 
             var lastHistoryEntryId = runtimeEntityType.AddProperty(
                 "LastHistoryEntryId",
-                typeof(ulong),
+                typeof(uint),
                 propertyInfo: typeof(ShikiUser).GetProperty("LastHistoryEntryId", BindingFlags.Public | BindingFlags.Instance | BindingFlags.DeclaredOnly),
                 fieldInfo: typeof(ShikiUser).GetField("<LastHistoryEntryId>k__BackingField", BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.DeclaredOnly));
